fix nasty re_entity_db <> re_query dep cycle
