+++ crates/re_entity_db/Cargo.toml
@@ -20,16 +20,18 @@ all-features = true
 default = []
 
 ## Enable (de)serialization using serde.
-serde = ["dep:serde", "dep:rmp-serde", "re_log_types/serde"]
+serde = ["dep:serde", "dep:rmp-serde", "re_log_types/serde", "re_query/serde"]
 
 
 [dependencies]
 re_data_store.workspace = true
 re_format.workspace = true
 re_int_histogram.workspace = true
+re_log.workspace = true
 re_log_encoding = { workspace = true, optional = true }
 re_log_types.workspace = true
-re_log.workspace = true
+re_query.workspace = true
+re_query_cache.workspace = true
 re_smart_channel.workspace = true
 re_tracing.workspace = true
 re_types_core.workspace = true
