+++ crates/re_query/src/lib.rs
@@ -8,7 +8,9 @@ mod util;
 pub use self::archetype_view::{ArchetypeView, ComponentWithInstances};
 pub use self::query::{get_component_with_instances, query_archetype};
 pub use self::range::range_archetype;
-pub use self::util::query_archetype_with_history;
+pub use self::util::{
+    query_archetype_with_history, ExtraQueryHistory, VisibleHistory, VisibleHistoryBoundary,
+};
 
 // Used for doc-tests
 #[doc(hidden)]
