+++ crates/re_query_cache/Cargo.toml
@@ -22,7 +22,6 @@ default = []
 [dependencies]
 # Rerun dependencies:
 re_data_store.workspace = true
-re_entity_db.workspace = true
 re_format.workspace = true
 re_log.workspace = true
 re_log_types.workspace = true
