+++ crates/re_query_cache/src/query.rs
@@ -2,8 +2,8 @@ use paste::paste;
 use seq_macro::seq;
 
 use re_data_store::{DataStore, LatestAtQuery, RangeQuery, TimeInt, TimeRange, Timeline};
-use re_entity_db::{ExtraQueryHistory, VisibleHistory};
 use re_log_types::{EntityPath, RowId};
+use re_query::{ExtraQueryHistory, VisibleHistory};
 use re_types_core::{components::InstanceKey, Archetype, Component};
 
 use crate::AnyQuery;
